fix: userId 타입 수정
user의 수가 많지 않다는 것을 가정해 userId를 Integer 타입으로 변경했습니다.

+++ src/main/java/com/cleanengine/coin/order/domain/BuyOrder.java
@@ -19,7 +19,7 @@ public class BuyOrder {
     private String ticker;
 
     @Column(name="user_id", nullable = false, updatable = false)
-    private Long userId;
+    private Integer userId;
 
     // TODO size를 VO로 바꾸어야 함
     @Column(name="size")
@@ -43,7 +43,7 @@ public class BuyOrder {
         this.size -= buySize;
     }
 
-    public static BuyOrder create(String ticker, Long userId, Double size, Double price, LocalDateTime createdAt,
+    public static BuyOrder create(String ticker, Integer userId, Double size, Double price, LocalDateTime createdAt,
                                   Boolean isMarketOrder, Boolean isBot) {
         BuyOrder buyOrder = new BuyOrder();
         buyOrder.ticker = ticker;
